Rename rule folder because it conflicts with Rule
