+++ src/rules/Type.js
