+++ src/rules/index.js
